created a support module to cut down on repetition in tests
