+++ spec/jquery_ui_rails_helpers/tabs_helper_spec.rb
@@ -1,26 +1,8 @@
 require "spec_helper"
 
 describe TabsHelper do
-  include ActionView::Helpers,
-          ActionView::Helpers::CaptureHelper,
-          ActionView::Helpers::JavaScriptHelper,
+  include ControllerTestHelpers,
           TabsHelper
-  
-  # allow tabs.create to run by stubbing an output_buffer
-  attr_accessor :output_buffer
-  @output_buffer = ""
-  
-  # stub content_for for testing
-  def content_for(name, content = nil, &block)
-    # this doesn't exist, and causes errors
-    @_content_for = {} unless defined? @_content_for
-    # we've got to initialize this, so we can concat to it
-    @_content_for[name] = '' if @_content_for[name].nil?
-    # now the rest is the same as in rails
-    content = capture(&block) if block_given?
-    @_content_for[name] << content if content
-    @_content_for[name] unless content
-  end
 
   it "should be empty, with an empty block" do
     output = ui_tabs do
